requirements.txt
dothttp-req<=0.0.43